Adiocionando a funcionalidade de reset de Login

+++ src/main/java/com/projeto/chamados/data/UsuarioEntity.java
@@ -14,7 +14,7 @@ public class UsuarioEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer loginid;
+    private Long loginid;
 
     @Email
     @NotBlank
