+++ src/main/java/com/projeto/chamados/repository/UsuarioRepository.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Repository;
  * @author heitor
  */
 @Repository 
-public interface UsuarioRepository extends  JpaRepository<UsuarioEntity,Integer>{
+public interface UsuarioRepository extends  JpaRepository<UsuarioEntity,Long>{
     
     boolean existsByEmail(String email);
     UsuarioEntity findByEmail(String email);
